requirements.txt
asyncpg==0.29.0
prometheus-client==0.20.0
PyYAML==6.0.2
pandas==2.2.2
matplotlib==3.9.2
psutil==6.0.0
rich==13.9.3
numpy==2.0.2


